requirements.txt
Pillow==8.1.1
tensorboardX==2.1
tensorboard==2.3.0
torch==1.10.0
torchvision==0.11.0
scipy==1.6.1
tqdm==4.64.0